+++ CHANGELOG.md
@@ -7,9 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2025-03-09
+
+### Added
+
+- Add `vscode-marketplace-client` dependency to check for extension updates and display a notification
+
+### Changed
+
+- Improve localization strings for the extension
+
 ## [1.0.0] - 2024-12-08
 
 - Initial release of Dart Barrel File extension
 
-[Unreleased]: https://github.com/ManuelGil/ovsx-dart-barrel/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/ManuelGil/ovsx-dart-barrel/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/ManuelGil/ovsx-dart-barrel/compare/v1.0.0...v1.1.0
 [1.0.0]: https://githttps://github.com/ManuelGil/ovsx-dart-barrel/releases/tag/v1.0.0